test: isolate SDK fixtures and simplify Windows shards
