+++ sdk/typescript/tests-ts/api.test.ts
@@ -3,7 +3,6 @@ import {
   copyFile,
   cp,
   mkdir,
-  mkdtemp,
   readFile,
   readdir,
   realpath,
@@ -16,7 +15,6 @@ import * as fsPromises from "node:fs/promises";
 import { execFileSync } from "node:child_process";
 import { createHash } from "node:crypto";
 import { existsSync } from "node:fs";
-import { tmpdir } from "node:os";
 import { basename, join } from "node:path";
 import { fileURLToPath, pathToFileURL } from "node:url";
 import { Codex, type CodexOptions, type ThreadEvent } from "@openai/codex-sdk";
@@ -45,16 +43,24 @@ import {
   type JsonObject,
 } from "../src/config.js";
 import { estimateScanCost, type ScanCost } from "../src/cost.js";
-import {
-  resolveCodexCommand,
-  runWorkbench,
-  setCodexSecurityCredentialLogout,
-} from "../src/runtime.js";
+import { resolveCodexCommand, runWorkbench } from "../src/runtime.js";
 import { normalizeTarget } from "../src/targets.js";
 import { SYNTHETIC_CREDENTIALS } from "./cli-fixtures.js";
 import { INTEGRATION_TARGET, PLUGIN_ROOT } from "./plugin-root.js";
-import { preparedRuntime } from "./support/api-events.js";
-import { runMockInSubprocess } from "./support/isolated-mock.js";
+import {
+  mockScanRegistration,
+  mockWorkbench,
+  shellEnvironmentReference,
+  SHELL_ENVIRONMENT_PREFIX,
+  TestClient,
+  TEST_SNAPSHOT_DIGEST,
+} from "./support/api-client.js";
+import {
+  completedEvents,
+  createApiTestFixtures,
+  preparedRuntime,
+} from "./support/api-events.js";
+import { runTestInSubprocess } from "./support/test-subprocess.js";
 
 type ScanObserverName = Parameters<
   NonNullable<ScanOptions["onObserverError"]>
@@ -62,13 +68,9 @@ type ScanObserverName = Parameters<
 
 const REPOSITORY_ROOT = fileURLToPath(new URL("../../..", import.meta.url));
 const EXAMPLE = join(PLUGIN_ROOT, "examples", "completed-scan");
-const temporaryDirectories: string[] = [];
-const TEST_SNAPSHOT_DIGEST = `codex-security-snapshot/v1:sha256:${"a".repeat(64)}`;
-const SHELL_ENVIRONMENT_PREFIX = process.platform === "win32" ? "$env:" : "$";
-
-function shellEnvironmentReference(name: string, suffix = ""): string {
-  return `"${SHELL_ENVIRONMENT_PREFIX}${name}${suffix}"`;
-}
+const { cleanup, copyCompletedScan, temporaryDirectory } =
+  createApiTestFixtures();
+afterEach(cleanup);
 
 const EXTERNAL_PROVIDER_CASES = [
   [
@@ -130,79 +132,6 @@ const BEDROCK_AUTHENTICATION_CASES = [
   ],
   ["default AWS credential chain", {}, "default_credential_chain"],
 ] as const;
-const TestClientBase = CodexSecurity as unknown as new (
-  config: Record<string, unknown>,
-  dependencies: Record<string, unknown>,
-) => CodexSecurity;
-
-function mockScanRegistration(args: readonly string[]) {
-  const recipe = JSON.parse(args[args.indexOf("--recipe-json") + 1]!) as {
-    repositoryRevision?: string;
-    target: { kind: string };
-  };
-  const kind =
-    recipe.target.kind === "refs" || recipe.target.kind === "working_tree"
-      ? "git_diff"
-      : recipe.repositoryRevision === undefined
-        ? "directory_snapshot"
-        : "git_revision";
-
-  return {
-    scanId: "scan_example_001",
-    targetId: "target_sha256_example",
-    targetRevision: recipe.repositoryRevision ?? "unversioned",
-    scanDir: args[args.indexOf("--scan-dir") + 1],
-    contract: {
-      target: {
-        allowedKinds: [kind],
-        ...(kind === "directory_snapshot"
-          ? { requiredSnapshotDigest: TEST_SNAPSHOT_DIGEST }
-          : {}),
-      },
-    },
-  };
-}
-
-class TestClient extends TestClientBase {
-  public constructor(
-    config: Record<string, unknown>,
-    dependencies: Record<string, unknown>,
-  ) {
-    super(config, {
-      runWorkbench: async (_options: unknown, args: readonly string[]) =>
-        mockWorkbench(args),
-      ...dependencies,
-    });
-  }
-}
-
-function mockWorkbench(args: readonly string[]) {
-  if (args[0] === "register-cli-scan") return mockScanRegistration(args);
-  if (args[0] === "get-scan-feedback") {
-    return {
-      scanId: "scan_example_001",
-      targetId: "target_sha256_example",
-      falsePositives: [],
-    };
-  }
-  return {};
-}
-
-afterEach(async () => {
-  await Promise.all(
-    temporaryDirectories
-      .splice(0)
-      .map((path) => rm(path, { recursive: true, force: true })),
-  );
-});
-
-async function temporaryDirectory(): Promise<string> {
-  const path = await realpath(
-    await mkdtemp(join(tmpdir(), "codex-security-api-")),
-  );
-  temporaryDirectories.push(path);
-  return path;
-}
 
 function nodeCodex(script: string): {
   command: { command: string };
@@ -216,13 +145,6 @@ function nodeCodex(script: string): {
     },
     environment: { NODE_OPTIONS: `--import=${pathToFileURL(script).href}` },
   };
-}
-
-async function copyCompletedScan(root: string): Promise<string> {
-  const scanDir = join(root, "scan");
-  await cp(EXAMPLE, scanDir, { recursive: true });
-  await writeFile(join(scanDir, "report.md"), "# Scan report\n");
-  return scanDir;
 }
 
 async function writeUsageSession(
@@ -255,25 +177,6 @@ async function writeUsageSession(
       "",
     ].join("\n"),
   );
-}
-
-async function* completedEvents(): AsyncGenerator<ThreadEvent> {
-  yield { type: "thread.started", thread_id: "thread-1" };
-  yield { type: "turn.started" };
-  yield {
-    type: "item.completed",
-    item: { id: "message-1", type: "agent_message", text: "scan complete" },
-  };
-  yield {
-    type: "turn.completed",
-    usage: {
-      input_tokens: 10,
-      cached_input_tokens: 2,
-      cache_write_input_tokens: 0,
-      output_tokens: 3,
-      reasoning_output_tokens: 1,
-    },
-  };
 }
 
 describe("CodexSecurity orchestration", () => {
@@ -322,7 +225,10 @@ describe("CodexSecurity orchestration", () => {
           prepareRuntime: async () => preparedRuntime(codexHome),
           resolvePluginPython: async () => "/managed/python",
           repositoryRevision: async () => null,
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             if (args[0] === "register-cli-scan") {
               recipe = JSON.parse(args[args.indexOf("--recipe-json") + 1]!);
             }
@@ -672,12 +578,13 @@ describe("CodexSecurity orchestration", () => {
     const repository = join(root, "repository");
     await mkdir(repository);
 
-    for (const codexOverrides of [
+    const invalidSettings: JsonObject[] = [
       { model: "" },
       { model: 42 },
       { model_reasoning_effort: "" },
       { model_reasoning_effort: false },
-    ]) {
+    ];
+    for (const codexOverrides of invalidSettings) {
       const client = new TestClient({ codexOverrides }, { environment: {} });
       await expect(client.preflight(repository)).rejects.toThrow(
         /model|reasoning effort/u,
@@ -822,10 +729,8 @@ describe("CodexSecurity orchestration", () => {
             codex_api_key: "synthetic-forwarded-codex-key",
           },
         }),
-        resolvePluginPython: async (options: {
-          environment?: Record<string, string | undefined>;
-        }) => {
-          pythonEnvironment = options.environment;
+        resolvePluginPython: async (options) => {
+          pythonEnvironment = options?.environment;
           return "/managed/python";
         },
         prepareOutputDir: async () => scanDir,
@@ -1139,7 +1044,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           if (args[0] === "register-cli-scan") {
             savedRecipe = JSON.parse(args[args.indexOf("--recipe-json") + 1]!);
           }
@@ -1426,7 +1334,10 @@ describe("CodexSecurity orchestration", () => {
         prepareRuntime: async () => preparedRuntime(codexHome),
         resolvePluginPython: async () => "/managed/python",
         repositoryRevision: async () => null,
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           if (args[0] === "get-scan-feedback") {
             return {
               scanId: "scan_example_001",
@@ -1784,6 +1695,13 @@ describe("CodexSecurity orchestration", () => {
   });
 
   test("keeps a relative repository stable if runtime initialization changes cwd", async () => {
+    if (
+      runTestInSubprocess(
+        fileURLToPath(import.meta.url),
+        "keeps a relative repository stable if runtime initialization changes cwd",
+      )
+    )
+      return;
     const root = await temporaryDirectory();
     const initial = join(root, "initial");
     const elsewhere = join(root, "elsewhere");
@@ -1869,7 +1787,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -2077,7 +1998,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           if (args[0] === "register-cli-scan") {
             return {
               ...mockScanRegistration(args),
@@ -2157,7 +2081,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           if (args[0] !== "register-cli-scan") {
             return {
               scanId: "scan_example_001",
@@ -2353,7 +2280,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) =>
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> =>
           args[0] === "register-cli-scan"
             ? {
                 ...mockScanRegistration(args),
@@ -2390,7 +2320,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args[0]!);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -2454,7 +2387,10 @@ describe("CodexSecurity orchestration", () => {
           resolvePluginPython: async () => "/managed/python",
           prepareOutputDir: async () => scanDir,
           repositoryRevision: async () => "deadbeef",
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             commands.push(args[0]!);
             return mockWorkbench(args);
           },
@@ -2506,7 +2442,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) =>
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> =>
           args[0] === "register-cli-scan"
             ? { ...mockScanRegistration(args), scopeFileCount: 4_207 }
             : mockWorkbench(args),
@@ -2582,7 +2521,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) =>
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> =>
           args[0] === "register-cli-scan"
             ? { ...mockScanRegistration(args), scopeFileCount: 1_258 }
             : mockWorkbench(args),
@@ -2704,7 +2646,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -2803,7 +2748,7 @@ describe("CodexSecurity orchestration", () => {
         prepareRuntime: async () => ({
           ...runtime,
           plugin: {
-            ...(runtime["plugin"] as Record<string, unknown>),
+            ...runtime.plugin,
             pluginRoot,
             marketplaceRoot: pluginRoot,
             installedRoot: pluginRoot,
@@ -2812,7 +2757,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args[0]!);
           return args[0] === "register-cli-scan"
             ? mockScanRegistration(args)
@@ -2853,7 +2801,10 @@ describe("CodexSecurity orchestration", () => {
           resolvePluginPython: async () => "/managed/python",
           prepareOutputDir: async () => scanDir,
           repositoryRevision: async () => "deadbeef",
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             if (args[0] === "register-cli-scan") {
               return { ...mockScanRegistration(args), scanId };
             }
@@ -2949,7 +2900,10 @@ describe("CodexSecurity orchestration", () => {
           resolvePluginPython: async () => "/managed/python",
           prepareOutputDir: async () => scanDir,
           repositoryRevision: async () => "deadbeef",
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             commands.push(args);
             if (args[0] === "get-scan-feedback") {
               return {
@@ -3078,7 +3032,10 @@ describe("CodexSecurity orchestration", () => {
           resolvePluginPython: async () => "/managed/python",
           prepareOutputDir: async () => scanDir,
           repositoryRevision: async () => "deadbeef",
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             if (args[0] === "register-cli-scan") {
               return mockScanRegistration(args);
             }
@@ -3139,7 +3096,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -3205,7 +3165,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           return mockWorkbench(args);
         },
@@ -3372,7 +3335,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -3501,7 +3467,10 @@ describe("CodexSecurity orchestration", () => {
           resolvePluginPython: async () => "/managed/python",
           prepareOutputDir: async () => scanDir,
           repositoryRevision: async () => "deadbeef",
-          runWorkbench: async (_options: unknown, args: readonly string[]) => {
+          runWorkbench: async (
+            _options: unknown,
+            args: readonly string[],
+          ): Promise<JsonObject> => {
             commands.push(args);
             if (args[0] !== "complete-budget-exhausted-scan") {
               return mockWorkbench(args);
@@ -3533,7 +3502,7 @@ describe("CodexSecurity orchestration", () => {
               .digest("hex");
             await writeFile(manifestPath, `${JSON.stringify(manifest)}\n`);
             return {
-              scan: { warnings: [args[args.indexOf("--message") + 1]] },
+              scan: { warnings: [args[args.indexOf("--message") + 1]!] },
             };
           },
           createCodex: () => ({
@@ -3633,7 +3602,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           commands.push(args);
           if (args[0] === "register-cli-scan") {
             return mockScanRegistration(args);
@@ -3717,7 +3689,10 @@ describe("CodexSecurity orchestration", () => {
         resolvePluginPython: async () => "/managed/python",
         prepareOutputDir: async () => scanDir,
         repositoryRevision: async () => "deadbeef",
-        runWorkbench: async (_options: unknown, args: readonly string[]) => {
+        runWorkbench: async (
+          _options: unknown,
+          args: readonly string[],
+        ): Promise<JsonObject> => {
           if (args[0] === "get-scan-feedback") {
             return {
               scanId: "scan_example_001",
@@ -3822,7 +3797,7 @@ describe("CodexSecurity orchestration", () => {
     const python = Bun.which("python3") ?? Bun.which("python");
     expect(python).not.toBeNull();
     const environment = {
-      PATH: process.env["PATH"],
+      PATH: process.env["PATH"] ?? "",
       CODEX_SECURITY_STATE_DIR: stateDirectory,
     };
     const commands: Array<readonly string[]> = [];
@@ -3840,7 +3815,7 @@ describe("CodexSecurity orchestration", () => {
         runWorkbench: async (
           options: Parameters<typeof runWorkbench>[0],
           args: readonly string[],
-        ) => {
+        ): Promise<JsonObject> => {
           commands.push(args);
           const result = await runWorkbench(options, args);
           if (args[0] === "fail-scan") {
@@ -3895,7 +3870,7 @@ describe("CodexSecurity orchestration", () => {
     const python = Bun.which("python3") ?? Bun.which("python");
     expect(python).not.toBeNull();
     const environment = {
-      PATH: process.env["PATH"],
+      PATH: process.env["PATH"] ?? "",
       CODEX_SECURITY_STATE_DIR: stateDirectory,
     };
     const commands: Array<readonly string[]> = [];
@@ -3918,7 +3893,7 @@ describe("CodexSecurity orchestration", () => {
         runWorkbench: async (
           options: Parameters<typeof runWorkbench>[0],
           args: readonly string[],
-        ) => {
+        ): Promise<JsonObject> => {
           commands.push(args);
           return await runWorkbench(options, args);
         },
@@ -4068,252 +4043,6 @@ describe("CodexSecurity orchestration", () => {
     await client.close();
   });
 
-  test("keeps a private preflight snapshot isolated from persistent credentials", async () => {
-    const root = await temporaryDirectory();
-    const repository = join(root, "repository");
-    const ambientHome = join(root, "ambient-codex-home");
-    const scanDir = join(root, "scan");
-    await mkdir(repository);
-    await mkdir(ambientHome);
-    await mkdir(scanDir, { mode: 0o700 });
-    await writeFile(join(ambientHome, "auth.json"), "{}\n");
-    const interpreter =
-      Bun.which("python3") ?? Bun.which("python") ?? Bun.which("py");
-    expect(interpreter).not.toBeNull();
-    let capturedConfigPath: string | undefined;
-    let capturedCodexHome: string | undefined;
-    const unrelatedProjects = Object.fromEntries(
-      Array.from({ length: 256 }, (_, index) => [
-        join(root, `unrelated-project-${index}`),
-        { trust_level: "untrusted" },
-      ]),
-    );
-    const client = new TestClient(
-      {
-        pluginPath: PLUGIN_ROOT,
-        codexOverrides: {
-          approval_policy: "never",
-          features: { goals: true },
-          projects: {
-            ...unrelatedProjects,
-            [repository]: { trust_level: "trusted" },
-          },
-          mcp_servers: {
-            private: {
-              command: "echo",
-              env: { PRIVATE_TOKEN: "RUNTIME_MCP_SECRET" },
-            },
-          },
-          shell_environment_policy: {
-            set: { PRIVATE_TOKEN: "RUNTIME_SHELL_SECRET" },
-          },
-          responses_api_metadata: {
-            request_trace: "preserve-configured-metadata",
-          },
-        },
-      },
-      {
-        environment: { CODEX_HOME: ambientHome },
-        resolvePluginPython: async () => interpreter!,
-        prepareOutputDir: async () => scanDir,
-        repositoryRevision: async () => "deadbeef",
-        createCodex: (options: CodexOptions) => ({
-          startThread: () => ({
-            id: null,
-            async runStreamed(input: string) {
-              const configPath = options.env?.["CODEX_SECURITY_CONFIG_PATH"];
-              const codexHome = options.env?.["CODEX_HOME"];
-              expect(typeof configPath).toBe("string");
-              expect(typeof codexHome).toBe("string");
-              capturedConfigPath = configPath;
-              capturedCodexHome = codexHome;
-              expect(configPath!.startsWith(`${codexHome!}/`)).toBe(false);
-              expect(
-                parseToml(
-                  await readFile(join(codexHome!, "config.toml"), "utf8"),
-                ),
-              ).toMatchObject({
-                approval_policy: "never",
-                permissions: {
-                  codex_security_scan: {
-                    filesystem: {
-                      ":root": "read",
-                      ":workspace_roots": "write",
-                      [join(ambientHome, "state", "plugins", "codex-security")]:
-                        "write",
-                      [join(
-                        ambientHome,
-                        "state",
-                        "plugins",
-                        "codex-security",
-                        "codex-home",
-                      )]: "read",
-                    },
-                  },
-                },
-              });
-              const codexConfig = await readFile(
-                join(codexHome!, "config.toml"),
-                "utf8",
-              );
-              expect(options.env?.["CODEX_SECURITY_SURFACE"]).toBe("sdk");
-              expect(codexConfig).not.toContain("model_reasoning_summary");
-              expect(codexConfig).not.toContain("show_raw_agent_reasoning");
-              expect(options.config).not.toHaveProperty("projects");
-              expect(options.config).not.toHaveProperty("permissions");
-              expect(options.config).toMatchObject({
-                default_permissions: "codex_security_scan",
-                allow_login_shell: false,
-                model_reasoning_summary: "detailed",
-                show_raw_agent_reasoning: true,
-                windows: { sandbox: "unelevated" },
-                mcp_servers: {
-                  private: {
-                    command: "echo",
-                    env: { PRIVATE_TOKEN: "RUNTIME_MCP_SECRET" },
-                  },
-                },
-                shell_environment_policy: {
-                  set: { PRIVATE_TOKEN: "RUNTIME_SHELL_SECRET" },
-                },
-                responses_api_metadata: {
-                  request_trace: "preserve-configured-metadata",
-                  codex_security_surface: "sdk",
-                },
-              });
-              if (process.platform !== "win32") {
-                expect((await stat(configPath!)).mode & 0o777).toBe(0o600);
-              }
-              const serialized = await readFile(configPath!, "utf8");
-              expect(serialized).not.toContain("RUNTIME_MCP_SECRET");
-              expect(serialized).not.toContain("RUNTIME_SHELL_SECRET");
-              expect(serialized).not.toContain("mcp_servers");
-              expect(serialized).not.toContain("shell_environment_policy");
-              expect(parseToml(serialized)).toMatchObject({
-                projects: {
-                  [repository]: { trust_level: "trusted" },
-                },
-              });
-              expect(input).toContain(
-                `--config ${shellEnvironmentReference("CODEX_SECURITY_CONFIG_PATH")}`,
-              );
-              expect(input).toContain("--effective-config");
-              const shellEnvironment = options.env as Record<string, string>;
-              const helper = execFileSync(
-                interpreter!,
-                [
-                  join(PLUGIN_ROOT, "scripts", "config_preflight.py"),
-                  "--skill",
-                  "security-scan",
-                  "--config",
-                  shellEnvironment["CODEX_SECURITY_CONFIG_PATH"]!,
-                  "--cwd",
-                  repository,
-                  "--multi-agent-runtime-owner",
-                  "native",
-                  "--multi-agent-runtime-version",
-                  "v2",
-                  "--multi-agent-session-cap",
-                  "12",
-                  "--multi-agent-runtime-provenance",
-                  "tool-surface",
-                  "--runtime-check",
-                  "delegation_available=true",
-                  "--runtime-check",
-                  "goal_tools_available=true",
-                  "--effective-config",
-                  "features.goals=true",
-                ],
-                {
-                  env: {
-                    PATH: process.env["PATH"],
-                    CODEX_HOME: join(root, "denied"),
-                  },
-                  encoding: "utf8",
-                },
-              );
-              const preflight = JSON.parse(helper) as Record<string, unknown>;
-              expect(preflight["status"]).toBe("ready");
-              expect(preflight["config_resolution"]).toBe("manual-layers");
-              expect(preflight["config_paths"]).toEqual([configPath]);
-              await copyCompletedScan(root);
-              const manifestPath = join(scanDir, "scan-manifest.json");
-              const manifest = JSON.parse(
-                await readFile(manifestPath, "utf8"),
-              ) as { scan: { producer: { version: string } } };
-              const pluginManifest = JSON.parse(
-                await readFile(
-                  join(PLUGIN_ROOT, ".codex-plugin", "plugin.json"),
-                  "utf8",
-                ),
-              ) as { version: string };
-              manifest.scan.producer.version = pluginManifest.version;
-              await writeFile(manifestPath, JSON.stringify(manifest));
-              return { events: completedEvents() };
-            },
-          }),
-        }),
-      },
-    );
-
-    try {
-      await client.run(repository);
-      expect(capturedConfigPath).toBeDefined();
-      expect(capturedCodexHome).toBeDefined();
-    } finally {
-      await client.close();
-    }
-    expect(existsSync(capturedConfigPath!)).toBe(false);
-    expect(capturedCodexHome).toBe(
-      join(ambientHome, "state", "plugins", "codex-security", "codex-home"),
-    );
-    expect(existsSync(capturedCodexHome!)).toBe(true);
-  });
-
-  test("reuses keyring-compatible credentials across separate scan clients", async () => {
-    const root = await temporaryDirectory();
-    const repository = join(root, "repository");
-    const ambientHome = join(root, "ambient-codex-home");
-    const stateDirectory = join(root, "state");
-    const credentialHome = join(stateDirectory, "codex-home");
-    const runtimeHomes: string[] = [];
-    await mkdir(repository);
-    await mkdir(ambientHome);
-    await writeFile(join(ambientHome, "auth.json"), "{}\n");
-
-    for (const index of [0, 1]) {
-      const scanDir = join(root, `scan-${index}`);
-      await mkdir(scanDir, { mode: 0o700 });
-      const client = new TestClient(
-        { pluginPath: PLUGIN_ROOT },
-        {
-          environment: {
-            CODEX_HOME: ambientHome,
-            CODEX_SECURITY_STATE_DIR: stateDirectory,
-          },
-          resolvePluginPython: async () => "/managed/python",
-          prepareOutputDir: async () => scanDir,
-          repositoryRevision: async () => "deadbeef",
-          createCodex: (options: CodexOptions) => {
-            runtimeHomes.push(options.env?.["CODEX_HOME"] ?? "");
-            throw new Error("persistent credential scan reached");
-          },
-        },
-      );
-
-      try {
-        await expect(client.run(repository)).rejects.toThrow(
-          "persistent credential scan reached",
-        );
-      } finally {
-        await client.close();
-      }
-      expect(existsSync(credentialHome)).toBe(true);
-    }
-
-    expect(runtimeHomes).toEqual([credentialHome, credentialHome]);
-  });
-
   test.each([
     ["OpenAI", undefined, "OPENAI_API_KEY", "gpt-5.6-sol", undefined],
     ...EXTERNAL_PROVIDER_CASES,
@@ -4346,7 +4075,7 @@ describe("CodexSecurity orchestration", () => {
               ? {}
               : {
                   model_provider: provider,
-                  model_providers: { [provider]: providerConfig },
+                  model_providers: { [provider]: providerConfig! },
                 }),
           },
         },
@@ -4428,10 +4157,12 @@ describe("CodexSecurity orchestration", () => {
     });
 
     const clients = await Promise.all(
-      [
-        ["OPENAI_API_KEY", "gpt-5.6-sol", undefined],
-        ["OPENROUTER_API_KEY", "anthropic/claude-sonnet-4.5", "openrouter"],
-      ].map(async ([apiKey, model, provider], index) => {
+      (
+        [
+          ["OPENAI_API_KEY", "gpt-5.6-sol", undefined],
+          ["OPENROUTER_API_KEY", "anthropic/claude-sonnet-4.5", "openrouter"],
+        ] as const
+      ).map(async ([apiKey, model, provider], index) => {
         const scanDir = join(root, `parallel-api-key-scan-${index}`);
         await mkdir(scanDir, { mode: 0o700 });
         return new TestClient(
@@ -4499,133 +4230,6 @@ describe("CodexSecurity orchestration", () => {
         });
       }
       expect(scansStarted).toBe(2);
-    } finally {
-      await Promise.all(clients.map(async (client) => await client.close()));
-    }
-  });
-
-  test("runs parallel ChatGPT scans with isolated mutable configuration", async () => {
-    const root = await temporaryDirectory();
-    const repository = join(root, "repository");
-    const ambientHome = join(root, "ambient-codex-home");
-    const stateDirectory = join(root, "state");
-    const credentialHome = join(stateDirectory, "codex-home");
-    await mkdir(repository);
-    await mkdir(ambientHome);
-    await writeFile(join(ambientHome, "auth.json"), "{}\n");
-    let activeScans = 0;
-    let maximumActiveScans = 0;
-    const deepScanConfigPaths = new Set<string>();
-    let releaseScans!: () => void;
-    const concurrentScans = new Promise<void>((resolve) => {
-      releaseScans = resolve;
-    });
-
-    const clients = await Promise.all(
-      [0, 1].map(async (index) => {
-        const scanDir = join(root, `parallel-scan-${index}`);
-        await mkdir(scanDir, { mode: 0o700 });
-        return new TestClient(
-          {
-            pluginPath: PLUGIN_ROOT,
-            codexOverrides: {
-              model: index === 0 ? "gpt-5.6-sol" : "gpt-5.6-terra",
-            },
-          },
-          {
-            environment: {
-              CODEX_HOME: ambientHome,
-              CODEX_SECURITY_STATE_DIR: stateDirectory,
-            },
-            resolvePluginPython: async () => "/managed/python",
-            prepareOutputDir: async () => scanDir,
-            repositoryRevision: async () => "deadbeef",
-            createCodex: (options: CodexOptions) => {
-              expect(options.env?.["CODEX_HOME"]).toBe(credentialHome);
-              const expectedModel =
-                index === 0 ? "gpt-5.6-sol" : "gpt-5.6-terra";
-              expect(options.config?.["model"]).toBe(expectedModel);
-              const deepScanConfigPath =
-                options.env?.["CODEX_SECURITY_DEEP_SCAN_CONFIG_PATH"];
-              expect(typeof deepScanConfigPath).toBe("string");
-              deepScanConfigPaths.add(deepScanConfigPath!);
-              return {
-                startThread: () => ({
-                  id: null,
-                  async runStreamed() {
-                    expect(
-                      existsSync(
-                        join(credentialHome, ".codex-security-scan.lock"),
-                      ),
-                    ).toBe(false);
-                    activeScans += 1;
-                    maximumActiveScans = Math.max(
-                      maximumActiveScans,
-                      activeScans,
-                    );
-                    if (activeScans === 2) releaseScans();
-                    try {
-                      const credentialConfig = parseToml(
-                        await readFile(
-                          join(credentialHome, "config.toml"),
-                          "utf8",
-                        ),
-                      );
-                      expect(credentialConfig["model"]).toBeUndefined();
-                      const before = parseToml(
-                        await readFile(deepScanConfigPath!, "utf8"),
-                      );
-                      expect(before["deep_scan"]).toMatchObject({
-                        workers: index + 2,
-                      });
-                      await Promise.race([
-                        concurrentScans,
-                        new Promise((resolve) => setTimeout(resolve, 5_000)),
-                      ]);
-                      const after = parseToml(
-                        await readFile(deepScanConfigPath!, "utf8"),
-                      );
-                      expect(after["deep_scan"]).toMatchObject({
-                        workers: index + 2,
-                      });
-                      throw new Error("parallel managed scan reached");
-                    } finally {
-                      activeScans -= 1;
-                    }
-                  },
-                }),
-              };
-            },
-          },
-        );
-      }),
-    );
-
-    try {
-      const results = await Promise.allSettled(
-        clients.map((client, index) =>
-          client.run(repository, { mode: "deep", workers: index + 2 }),
-        ),
-      );
-      for (const result of results) {
-        expect(result).toMatchObject({
-          status: "rejected",
-          reason: expect.objectContaining({
-            message: "parallel managed scan reached",
-          }),
-        });
-      }
-      expect(existsSync(credentialHome)).toBe(true);
-      expect(maximumActiveScans).toBe(2);
-      expect(deepScanConfigPaths.size).toBe(2);
-      const pluginConfiguration = JSON.parse(
-        await readFile(join(PLUGIN_ROOT, ".mcp.json"), "utf8"),
-      ) as { mcpServers: Record<string, { env_vars: string[] }> };
-      expect(
-        pluginConfiguration.mcpServers["codex-security"]?.env_vars.includes(
-          "CODEX_SECURITY_DEEP_SCAN_CONFIG_PATH",
-        ),
-      ).toBe(true);
     } finally {
       await Promise.all(clients.map(async (client) => await client.close()));
     }
@@ -4715,10 +4319,7 @@ describe("CodexSecurity orchestration", () => {
         prepareRuntime: async () => ({
           ...preparedRuntime(codexHome),
           plugin: {
-            ...(preparedRuntime(codexHome)["plugin"] as Record<
-              string,
-              unknown
-            >),
+            ...preparedRuntime(codexHome).plugin,
             pluginRoot,
             installedRoot: pluginRoot,
           },
@@ -4831,7 +4432,7 @@ describe("CodexSecurity orchestration", () => {
           return {
             ...runtime,
             plugin: {
-              ...(runtime["plugin"] as Record<string, unknown>),
+              ...runtime.plugin,
               installedRoot: join(
                 codexHome,
                 "plugins",
@@ -4955,7 +4556,7 @@ describe("CodexSecurity orchestration", () => {
         {
           cwd: root,
           env: {
-            PATH: process.env["PATH"],
+            PATH: process.env["PATH"] ?? "",
             HOME: process.env["HOME"],
             ...environment,
             CODEX_SECURITY_TARGET_PATHS_FILE: capturedTargetPathsFile,
@@ -5539,12 +5140,9 @@ process.exit(2);
           credentialsAvailable: false,
         }),
         resolveCodexCommand: () => fakeCommand.command,
-        resolvePluginPython: async (options: {
-          environment?: Record<string, string | undefined>;
-          protectedRoot?: string;
-        }) => {
-          pythonEnvironment = options.environment;
-          pythonProtectedRoot = options.protectedRoot;
+        resolvePluginPython: async (options) => {
+          pythonEnvironment = options?.environment;
+          pythonProtectedRoot = options?.protectedRoot;
           return "/managed/python";
         },
         prepareOutputDir: async () => scanDir,
@@ -5649,69 +5247,6 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
     }
   });
 
-  test("reuses the managed runtime when scan authentication changes", async () => {
-    const root = await temporaryDirectory();
-    const repository = join(root, "repository");
-    const ambientHome = join(root, "ambient-codex-home");
-    const stateDirectory = join(root, "state");
-    const dedicatedHome = join(stateDirectory, "codex-home");
-    const scanDir = join(root, "scan");
-    const ambientAuthentication = '{"auth_mode":"chatgpt"}\n';
-    await mkdir(repository);
-    await mkdir(ambientHome);
-    await mkdir(scanDir, { mode: 0o700 });
-    await writeFile(join(ambientHome, "auth.json"), ambientAuthentication);
-    const runs: Array<{ home: string; apiKey?: string }> = [];
-    const client = new TestClient(
-      { pluginPath: PLUGIN_ROOT },
-      {
-        environment: {
-          CODEX_HOME: ambientHome,
-          CODEX_SECURITY_STATE_DIR: stateDirectory,
-          OPENAI_API_KEY: "synthetic-transient-key",
-        },
-        resolvePluginPython: async () => "/managed/python",
-        prepareOutputDir: async () => scanDir,
-        repositoryRevision: async () => "deadbeef",
-        createCodex: (options: CodexOptions) => {
-          runs.push({
-            home: options.env?.["CODEX_HOME"] ?? "",
-            ...(options.apiKey === undefined ? {} : { apiKey: options.apiKey }),
-          });
-          throw new Error("authentication-selected scan reached");
-        },
-      },
-    );
-
-    try {
-      await expect(client.run(repository, { auth: "api-key" })).rejects.toThrow(
-        "authentication-selected scan reached",
-      );
-      expect(runs[0]?.home).toBe(dedicatedHome);
-      expect(runs[0]?.apiKey).toBe("synthetic-transient-key");
-      expect(existsSync(join(dedicatedHome, "auth.json"))).toBe(false);
-
-      await expect(client.run(repository, { auth: "chatgpt" })).rejects.toThrow(
-        "authentication-selected scan reached",
-      );
-      expect(runs[1]).toEqual({ home: dedicatedHome });
-      expect(await readFile(join(dedicatedHome, "auth.json"), "utf8")).toBe(
-        ambientAuthentication,
-      );
-      await expect(client.run(repository, { auth: "api-key" })).rejects.toThrow(
-        "authentication-selected scan reached",
-      );
-      expect(runs[2]?.home).toBe(dedicatedHome);
-      expect(runs[2]?.apiKey).toBe("synthetic-transient-key");
-      expect(await readFile(join(dedicatedHome, "auth.json"), "utf8")).toBe(
-        ambientAuthentication,
-      );
-    } finally {
-      await client.close();
-    }
-    expect(existsSync(dedicatedHome)).toBe(true);
-  });
-
   test("does not cache an environment key as reusable file authentication", async () => {
     let imported = false;
     await expect(
@@ -5764,35 +5299,6 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
     expect(await readFile(join(credentialHome, "auth.json"), "utf8")).toBe(
       '{"token":"explicit"}\n',
     );
-  });
-
-  test("does not reimport ambient credentials after an explicit logout", async () => {
-    const root = await temporaryDirectory();
-    const ambientHome = join(root, "ambient-home");
-    const credentialHome = join(root, "credential-home");
-    await mkdir(ambientHome);
-    await mkdir(credentialHome, { mode: 0o700 });
-    await writeFile(join(ambientHome, "auth.json"), '{"token":"ambient"}\n');
-    await setCodexSecurityCredentialLogout(credentialHome, true);
-    let imported = false;
-
-    await expect(
-      initialCredentialsAvailable({}, ambientHome, credentialHome, async () => {
-        imported = true;
-        return true;
-      }),
-    ).resolves.toBe(false);
-    expect(imported).toBe(false);
-
-    await setCodexSecurityCredentialLogout(credentialHome, false);
-    await expect(
-      initialCredentialsAvailable(
-        {},
-        ambientHome,
-        credentialHome,
-        async () => true,
-      ),
-    ).resolves.toBe(true);
   });
 
   test("restores stored ChatGPT credentials when switching from an API-key scan", async () => {
@@ -5956,14 +5462,16 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
     const codexHome = join(root, "codex-home");
     await mkdir(repository);
     await mkdir(codexHome);
-    let releaseRuntime!: (runtime: Record<string, unknown>) => void;
+    let releaseRuntime!: (runtime: ReturnType<typeof preparedRuntime>) => void;
     let preparationStarted!: () => void;
     const started = new Promise<void>((resolve) => {
       preparationStarted = resolve;
     });
-    const prepared = new Promise<Record<string, unknown>>((resolve) => {
-      releaseRuntime = resolve;
-    });
+    const prepared = new Promise<ReturnType<typeof preparedRuntime>>(
+      (resolve) => {
+        releaseRuntime = resolve;
+      },
+    );
     let createCodexCalled = false;
     const client = new TestClient(
       {},
@@ -6008,14 +5516,16 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
     await mkdir(repository);
     await mkdir(codexHome);
     await mkdir(scanDir, { mode: 0o700 });
-    let releaseRuntime!: (runtime: Record<string, unknown>) => void;
+    let releaseRuntime!: (runtime: ReturnType<typeof preparedRuntime>) => void;
     let preparationStarted!: () => void;
     const started = new Promise<void>((resolve) => {
       preparationStarted = resolve;
     });
-    const prepared = new Promise<Record<string, unknown>>((resolve) => {
-      releaseRuntime = resolve;
-    });
+    const prepared = new Promise<ReturnType<typeof preparedRuntime>>(
+      (resolve) => {
+        releaseRuntime = resolve;
+      },
+    );
     const client = new TestClient(
       {},
       {
@@ -6216,7 +5726,7 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
 
   test("cleans the bootstrap workspace when credential-home cleanup fails", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "cleans the bootstrap workspace when credential-home cleanup fails",
       )
@@ -6279,7 +5789,7 @@ if ([basename(process.argv[1]), ...process.argv.slice(2)].join(" ") !== "login s
 
   test("attempts both preparation cleanups and preserves the preparation and cleanup failures", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "attempts both preparation cleanups and preserves the preparation and cleanup failures",
       )
