+++ sdk/typescript/tests-ts/runtime.test.ts
@@ -27,7 +27,6 @@ import {
   relative,
   sep,
 } from "node:path";
-import { fileURLToPath, pathToFileURL } from "node:url";
 import { promisify } from "node:util";
 import { brotliDecompressSync } from "node:zlib";
 import { afterEach, describe, expect, mock, test } from "bun:test";
@@ -74,7 +73,7 @@ import {
   verifyStableWindowsCredentialDescendants,
 } from "../src/runtime.js";
 import { loadBundledRuntime, PLUGIN_ROOT } from "./plugin-root.js";
-import { runMockInSubprocess } from "./support/isolated-mock.js";
+import { runTestInSubprocess } from "./support/test-subprocess.js";
 
 const temporaryDirectories: string[] = [];
 const testPosix = process.platform === "win32" ? test.skip : test;
@@ -890,7 +889,7 @@ describe("plugin runtime preparation", () => {
 
   test("cancels configured plugin directory discovery", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "cancels configured plugin directory discovery",
       )
@@ -1014,7 +1013,7 @@ describe("plugin runtime preparation", () => {
     "rejects a queued plugin directory replaced with a symlink",
     async () => {
       if (
-        runMockInSubprocess(
+        runTestInSubprocess(
           import.meta.path,
           "rejects a queued plugin directory replaced with a symlink",
         )
@@ -1327,7 +1326,7 @@ describe("plugin runtime preparation", () => {
 
   test("imports ambient auth when credential files do not support hard links", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "imports ambient auth when credential files do not support hard links",
       )
@@ -2259,7 +2258,7 @@ describe("runtime directories and plugin Python boundary", () => {
 
   test("retries Windows credential verification when a descendant disappears", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "retries Windows credential verification when a descendant disappears",
       )
@@ -2304,7 +2303,7 @@ describe("runtime directories and plugin Python boundary", () => {
 
   test("rejects Windows credential descendants that repeatedly disappear", async () => {
     if (
-      runMockInSubprocess(
+      runTestInSubprocess(
         import.meta.path,
         "rejects Windows credential descendants that repeatedly disappear",
       )
@@ -3359,193 +3358,6 @@ describe("runtime directories and plugin Python boundary", () => {
       expect(JSON.parse(inspection.stdout)).toEqual({ unexpected: 0 });
       expect(await readFile(auth, "utf8")).toContain("synthetic-root");
       expect(await readFile(nestedAuth, "utf8")).toContain("synthetic-nested");
-    },
-  );
-
-  test.skipIf(
-    process.platform !== "win32" ||
-      process.env["GITHUB_ACTIONS"] !== "true" ||
-      process.env["RUNNER_ENVIRONMENT"] !== "github-hosted" ||
-      process.env["CODEX_SECURITY_ALLOW_MACHINE_POLICY_TEST"] !== "true",
-  )(
-    "prepares managed credential homes under constrained PowerShell",
-    async () => {
-      const root = await temporaryDirectory();
-      const powershell = join(
-        process.env["SystemRoot"] ?? "C:\\Windows",
-        "System32",
-        "WindowsPowerShell",
-        "v1.0",
-        "powershell.exe",
-      );
-      const constrainedEnvironment = {
-        ...process.env,
-        CODEX_SECURITY_STATE_DIR: join(root, "state"),
-        PSModulePath: join(root, "untrusted-or-incompatible-modules"),
-        PSMODULEPATH: join(root, "uppercase-untrusted-modules"),
-      };
-      const registry = join(dirname(dirname(dirname(powershell))), "reg.exe");
-      const policyKey =
-        "HKLM\\SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment";
-      const policyName = "__PSLockdownPolicy";
-      const original = spawnSync(
-        registry,
-        ["query", policyKey, "/v", policyName],
-        { encoding: "utf8", timeout: 15_000, windowsHide: true },
-      );
-      expect(original.status === 0 || original.status === 1).toBe(true);
-      const originalEntry =
-        original.status === 0
-          ? /^\s*__PSLockdownPolicy\s+(REG_[A-Z_]+)\s*(.*?)\s*$/mu.exec(
-              original.stdout,
-            )
-          : null;
-      if (original.status === 0) expect(originalEntry).not.toBeNull();
-      const originalPolicy =
-        originalEntry === null
-          ? null
-          : { type: originalEntry[1]!, value: originalEntry[2]! };
-      const enabled = spawnSync(
-        registry,
-        ["add", policyKey, "/v", policyName, "/t", "REG_SZ", "/d", "4", "/f"],
-        { encoding: "utf8", timeout: 15_000, windowsHide: true },
-      );
-      expect(enabled.status).toBe(0);
-
-      try {
-        const mode = spawnSync(
-          powershell,
-          [
-            "-NoLogo",
-            "-NoProfile",
-            "-NonInteractive",
-            "-Command",
-            "$ExecutionContext.SessionState.LanguageMode",
-          ],
-          {
-            encoding: "utf8",
-            env: constrainedEnvironment,
-            timeout: 15_000,
-            windowsHide: true,
-          },
-        );
-        expect(mode.status).toBe(0);
-        expect(mode.stdout.trim()).toBe("ConstrainedLanguage");
-
-        const oldImplementation = spawnSync(
-          powershell,
-          [
-            "-NoLogo",
-            "-NoProfile",
-            "-NonInteractive",
-            "-Command",
-            "$ErrorActionPreference = 'Stop'; New-Object System.Security.AccessControl.DirectorySecurity",
-          ],
-          {
-            encoding: "utf8",
-            env: constrainedEnvironment,
-            timeout: 15_000,
-            windowsHide: true,
-          },
-        );
-        expect(oldImplementation.status).not.toBe(0);
-
-        const trustedPowerShellEnvironment = {
-          ...Object.fromEntries(
-            Object.entries(process.env).filter(
-              ([name]) => name.toUpperCase() !== "PSMODULEPATH",
-            ),
-          ),
-          PSModulePath: join(dirname(powershell), "Modules"),
-        };
-        const guest = spawnSync(
-          powershell,
-          [
-            "-NoLogo",
-            "-NoProfile",
-            "-NonInteractive",
-            "-Command",
-            [
-              "Microsoft.PowerShell.Utility\\ConvertFrom-SddlString -Sddl 'O:LGG:SYD:(A;;GA;;;SY)'",
-              "Microsoft.PowerShell.Utility\\Select-Object -ExpandProperty RawDescriptor",
-              "Microsoft.PowerShell.Utility\\Select-Object -ExpandProperty Owner",
-              "Microsoft.PowerShell.Utility\\Select-Object -ExpandProperty Value",
-            ].join(" | "),
-          ],
-          {
-            encoding: "utf8",
-            env: trustedPowerShellEnvironment,
-            timeout: 15_000,
-            windowsHide: true,
-          },
-        );
-        expect(guest.status).toBe(0);
-        expect(guest.stdout.trim()).toMatch(/^S-1-(?:\d+-)*501$/u);
-        const home = join(root, "state", "codex-home");
-        await mkdir(home, { recursive: true });
-        const foreignGrant = spawnSync(
-          join(dirname(dirname(dirname(powershell))), "icacls.exe"),
-          [home, "/grant", `*${guest.stdout.trim()}:(OI)(CI)R`],
-          { encoding: "utf8", timeout: 15_000, windowsHide: true },
-        );
-        expect(foreignGrant.status).toBe(0);
-
-        const fixtureModule = join(root, "runtime-node-fixture.mjs");
-        const build = spawnSync(
-          process.execPath,
-          [
-            "build",
-            fileURLToPath(new URL("../src/runtime.ts", import.meta.url)),
-            "--target=node",
-            "--format=esm",
-            `--outfile=${fixtureModule}`,
-          ],
-          { encoding: "utf8", timeout: 30_000, windowsHide: true },
-        );
-        expect(build.status).toBe(0);
-        const selectedNode = spawnSync("node", ["-p", "process.execPath"], {
-          encoding: "utf8",
-          timeout: 15_000,
-          windowsHide: true,
-        });
-        expect(selectedNode.status).toBe(0);
-        const fixture = spawnSync(
-          selectedNode.stdout.trim(),
-          [
-            "--input-type=module",
-            "--eval",
-            `import { prepareCodexSecurityCredentialHome } from ${JSON.stringify(pathToFileURL(fixtureModule).href)}; await prepareCodexSecurityCredentialHome();`,
-          ],
-          {
-            encoding: "utf8",
-            env: constrainedEnvironment,
-            timeout: 30_000,
-            windowsHide: true,
-          },
-        );
-        expect(fixture.stderr).toBe("");
-        expect(fixture.status).toBe(0);
-        expect(existsSync(home)).toBe(true);
-      } finally {
-        const restore = spawnSync(
-          registry,
-          originalPolicy === null
-            ? ["delete", policyKey, "/v", policyName, "/f"]
-            : [
-                "add",
-                policyKey,
-                "/v",
-                policyName,
-                "/t",
-                originalPolicy.type,
-                "/d",
-                originalPolicy.value,
-                "/f",
-              ],
-          { encoding: "utf8", timeout: 15_000, windowsHide: true },
-        );
-        expect(restore.status).toBe(0);
-      }
     },
   );
